requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart
email-validator
httpx
pytest
pyodbc
pymysql
python-dotenv